updated the op-code suproutines movem.l a1 as well and to avoid using lea
I attempted changing STRING_FROM_WORD, but that didn't turn out nicely. You would have to move a long from the register to (a1). If a1 is currently odd, this causes errors.

You would have to move byte by byte. But wait, it has to be the most significant bytes first. This complicates things unnecessarily.

I know I'm the one who suggested this idea, but having STRING_FROM_WORD store the string value in a register is not a good idea.