Implemented AND and GET_EA
GET_EA is a subroutine that will hanldle all the effective address part of the op-code if it is used correctly.

It is a powerful subroutine, but use it with caution and awareness. While you can usually blug and chug, that won't always be the case.

I implemented another utility subroutine GET_LIGHT_PURPLE_SIZE that handles printing the size and setting d3.b for you. Use it in conjunction with GET_EA. If your op-code doesn't have a light purple size, that is a good indicator you need to be careful with GET_EA.