Implemented and tested MOVE and MOVEM_TESTS
With this the project is complete. However, I need to review and test MOVE and MOVEM more because they are easy to make mistakes in.

After that, I need to combine all the tests and make a thorough and complete test file. I will store the output of this test file.

Then I must review, clean, and refactor the code. After I refactor it, I will get its output and diff it with the output before refactoring.

After that, it is documentation.